Dragging a non-bool predicate from the catalog now makes a compound node
Non-Boolean predicates now produce a Compound Action node (with an 
action that calls the method of the predicate) when dragged from the 
catalog; Note that other changes (e.g. changing the graphical
representation of the dragged predicate from a condition node to a
compound action node) are probably needed to fully address this feature
request.