Major overhaul to SB_Autocomplete
The main change is a deduplication of autocomplete features; before this
commit, SB_Autocomplete duplicated much of the functionality of
AutoCompleteSimbionicProjectDefinitions. This deduplication is intended
to make
fixes to autocompletion (Currently issue #8) easier.

Currently, SB_Autocomplete does include verbatum duplication of
SB_AutocompleteTextArea behavior, though that duplication is likely not
an
issue for changing autocomplete behavior.