Significant overhaul of match handling by autocomplete
The main change was the introduction of SB_Auto_Match (currently an
inner class
of AutoCompletionHelper), a class that abstractly represents an
autocompletion
match. The most relevant abstraction for this commit is that a match can
provide separate strings for adding text to an editable area (e.g.
SB_Autocomplete, SB_AutocompleteTextArea) and for rendering in the
pop-up pane
containing a list of options. These strings are, respectively, the
"insertion"
and "display" strings.

Previously, the "display" string was what was used to represent a
"match". This
meant that inserting a different string during autocompletion required
parsing.

Unfortunately, a existing functionality appears to use the "display"
string as
the canonical representation of SB_Variable (locals, globals, constants,
parameters) and SB_Function (behaviors, actions, predicates) instances
for the
purpose of autocomplete. (Note that this is already flawed, as a
SB_Function
with nonzero parameters may be transformed into more than one display
string
depending on user input!)

The majority of changes are in AutoCompletionHelper and
AutoCompleteSimbionicProjectDefinitions, the main classes for handing
matches.
Some cascading changes were needed to provide Strings to methods that
previously directly accessed the list of "display" strings maintained by
AutoCompleteSimbionicProjectDefinitions.

This commit attempts to preserve existing functionality by using the
"display"
string for many match-management purposes. However, this may limit
future
changes. No new behavior is intended.

As of this commit, SB_Auto_Match objects are a "pair" containing a
pregenerated
"insertion" and "display" string. Future changes may be appropriate for
further
customization of "display" rendering.