Folders for Globals and Constants
The main content of this commit builds on earlier XML changes to add
folders for globals and constants in the SimBionic project file editor.
Note that a number of these changes are fixes for behavior that was
untested as of the previous commit.

A few changes in this commit change the editor's behavior to create UI
folders in "default order" (A lexical sort where folders come before
non-folders but each category is alphabetically sorted) when loading a
project; the previous behavior would load all "top-level" items (e.g.
folders and behaviors directly under the "Behaviors" tree) in the order
encountered in the project file.