Introduction of Java class name shortening
This was a "bonus goal" for "Fix Intellisense" (OH SIM #8)": using
unqualified names by default, only using the longer qualified names
when unqualified names introduce ambiguity (e.g. shorten java.awt.List
to List unless java.util.List is also known to the project.)

The name shortening of this commit is overeager; unknown  classes are
assumed to be non-colliding.

The introduction of "canonical name" to SB_Auto_Match and the changes
to SB_Autocomplete and SB_AutocompleteTextArea to use canonical names
is future-proofing in case _filterByType is reenabled.