Test updates for Constant and Global Folders
SimBionicJavaReadWriteTest: Updated to place constants and globals in
folders, and verify that the read structure is as expected.

Constants.sbj: Some constants moved to folders to verify that folders
do not interfere with the resolution of constants.

Also included are a few notes on the additional changes to the autogen
of XML readers and writers to use the new folder models.