Introduction of SB_BehaviorSummaryPane
SB_BehaviorSummaryPane displays the name and description of a behavior;
in the current UI layout (handled by changes in SimBionicFrame), this
display appears above the behavior editing area.

Additional changes to EditorTree and SB_Canvas are an attempt to ensure
that the summary pane gets updated when the current behavior is renamed
or has its description changed. Note that the implementation uses
ComponentRegistry in a way that may be difficult to refactor in the
future. Underlying issues with behavior navigation may remain.