SB_BindingsPanel and related updates
Changes in the SB_BindingsPanel now "write through" to the
SB_BindingsHolder being edited. The OK button is almost redundant (it
will still be used by SB_ToolBar until the bindings popup is fully
removed); the Cancel button no longer does what it is supposed to do.
Every bindings modification is now put on the undo stack.

Undo updates the NodeEditorPane as well as the underlying data;
howerver, this happens because of an implementation detail for undo.
Future investigation may be needed if the undo system is updated.

The choice of AbstractTableModel#fireTableDataChanged() was one of
convenience, as the custom listeners I wrote do not use the details of
the fired events. More appropriate AbstractTableModel methods include
firTableRowsInserted, etc.

Bug fix:
Changed SB_BindingsTable#getBindingsCopy to use copyBindings to make
a deep copy instead of a shallow copy; it turned out that the bindings
in the list are assumed to not be shared with the SB_BindingsHolder
being edited (Manifestation: the display string of SB_Rectangle and
SB_Condition objects would fail to update)