Removal of the dialog for editing bindings
This commit removes the code that displays a dialog for editing the
bindings of SB_Element and SB_Connector instances (with the exception
of SB_MultiRectangle). The NodeEditorPane and its write-through editor
are the indended replacement of this functionality.

Part of this removal includes the removal of automatically adding a new
binding to a SB_BindingsHolder instance through certain codepaths.

Unfortunately, the dialog for editing compound actions remains, as it
is implemented with its own classes, which will need aditional
modification.