Initial commit of compound action editing in NEP
This commit introduces the SB_MultiBindingsEditor for the
NodeEditorPane; the dialog for editing compound actions remains
untouched, and will probably remain so until it is removed.

Included is an overambitious refactoring of SB_BindingsPanel into the
class AbstractBindingsEditor in an attempt to reuse most of the
SB_BindingsPanel functionality for SB_MultiBindingsEditor.

Also included is a UI proof-of-concept of switching the NodeEditorPane
display from an editor for bindings and expressions and an editor for
compound actions.