Removal of unnecessary null checks for _glassPane
SB_Autocomplete and SB_AutocompleteTextArea run null checks on
their _glassPane fields, and sometimes attempt alternative actions
when the field is null. However, since the original commit of SimBionic
to GitHub, no code attempted to set this field to null.

Moreover, the alternate paths when _glassPane is null are the only
paths that call some methods of SB_AutocompleteListener. This change
may motivate further cleanup of SB_AutocompleteListener.

Two of the methods implemented by SB_ToolBar appear to be part of an
attempt to manage a SB_ToolBar-specific autocomplete implementation
that is both unused and missing features implemented for autocomplete
with SB_GlassPane.