Initial commit of expression editor for the NEP
The expression editor writes changes when the expression autocomplete
text area permanently loses focus to something that is not the
SB_GlassPane managed by that SB_AutocompleteTextArea instance. Like the
implementation used for the comment editor, this allows for limited
"write through" behavior.

The specifics of the focus-based "editing finished" detection for the
SB_AutocompleteTextArea seem somewhat unpleasant and should probably be
replaced if better alternatives are available.