Fixed NEP to handle changing displayed behavior
The node editor panel now writes changes when switching between
displayed behaviors.
Thee editor panel assumes that the highlighted node of the current
behavior is the one that should be displayed; this means that switching
to a behavior with a highlighted node will show the highlighted node in
the node editor panel.

Note that SB_TabbedCanvas#setNewPolyForCanvas was a stop-gap measure to
avoid replacing the 150+ direct accesses to SB_Canvas#_poly with
calls to getters and setters; this might leave loopholes for changing
behaviors without triggering the "behavior changed" behavior of the
node editor panel.

Also included is an attempt to clean up the Node Editor Panel and
SimBionicFrame classes.